backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
feedparser==6.0.11
requests==2.32.3
google-api-python-client==2.159.0
google-auth-oauthlib==1.2.1
python-dotenv==1.0.1
APScheduler==3.10.4
